requirements.txt
# Core dependencies
aiohttp>=3.9.0
python-socketio[asyncio_client]>=5.10.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# AI Providers
aisuite>=0.1.0
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0

# Music APIs
discogs-client>=2.3.0
spotipy>=2.23.0

# Content moderation (optional - has dependency issues)
# profanity-filter>=1.3.3
better-profanity>=0.7.0

# Audio processing (optional for genre classification)
librosa>=0.10.0
numpy>=1.24.0

# Caching
cachetools>=5.3.0

# Utilities
asyncio>=3.4.3

